feat: add function to play a entire game
a game consists of 5 rounds, each round is played
with the function "playRound". This function is moved
into the playGame-function to get access to the
variables "humanScore" and "computerScore" which are
defined in the "playGame"-function.

A function "declareWinner" is helping with declaring
the final winner of 5 rounds. It is possible to end
with a tie.

On the end of the "playGame"-function the variables
"humanScore" and "computerScore" are set back to 0,
if we want to start a new game;