refactor: playRound-function changed for ui-use
Changed the playRound-function to only get one argument with the
players choice. The computer choice will be generated in the
function, as soon as the user clicks the button to submit his choice.

the variables for humanScore and computerScore are moved outside the
playGame-function, to use them with the playRound-function.